Abstract file reading and support jekyll frontmatter
Jekyll is a popular static blogging engine and frontmatter is
distinguished by a `---` prefix and suffix at the beginning of the file.
This commit originally set out to add support to pagenode for files with
this more traditional style of frontmatter. In doing so, it was
discovered that the mark up files were read in two separate places, so
this adds a new `PN_FileReader` class to abstract file reading into a
single location.